+++ Code/Engine/Foundation/IO/Implementation/BinaryStreamOperationsOther_inl.h
@@ -2,7 +2,15 @@
 #pragma once
 
 /// \brief Operator to serialize ezIAllocator::Stats objects.
-void EZ_FOUNDATION_DLL operator<< (ezIBinaryStreamWriter& Stream, const ezIAllocator::Stats& rhs);
+EZ_FOUNDATION_DLL void operator<< (ezIBinaryStreamWriter& Stream, const ezIAllocator::Stats& rhs);
 
 /// \brief Operator to serialize ezIAllocator::Stats objects.
-void EZ_FOUNDATION_DLL operator>> (ezIBinaryStreamReader& Stream, ezIAllocator::Stats& rhs);
+EZ_FOUNDATION_DLL void operator>> (ezIBinaryStreamReader& Stream, ezIAllocator::Stats& rhs);
+
+class ezTime;
+
+/// \brief Operator to serialize ezTime objects.
+EZ_FOUNDATION_DLL void operator<< (ezIBinaryStreamWriter& Stream, ezTime Value);
+
+/// \brief Operator to serialize ezTime objects.
+EZ_FOUNDATION_DLL void operator>> (ezIBinaryStreamReader& Stream, ezTime& Value);